Goodbye, Github 02 (Finale)
GitHub has provided many things, however it lacks the ability to privatize my own code. This issue has become a very large issue with the SBC (i hope all of you fucking die.), I have filed DMCA takedowns with evidence, and they should be decimated within the next week or so. Thanks to the GitHub legal team and the users. I doubt anyone is reading this, but the switch to private hosting has alr happened, and shouldn't impact user experience, just what i have to do to update it.